Introduce `EmberCli::Sprockets`
Manage integration with Asset Pipeline with `EmberCli::Sprockets`.

This commit moves the focus of the Rails helpers' integration from the
`EmberCli::App` instance to the `EmberCli::Sprockets` instance.

This commit further simplifies and reduces the responsibilities of the
`App` class.
