+++ app/helpers/ember_rails_helper.rb
@@ -6,20 +6,14 @@ module EmberRailsHelper
 
     head, body = markup_capturer.capture
 
-    html = EmberCli[name].index_html(
-      sprockets: self,
-      head: head,
-      body: body,
-    )
-
-    render inline: html
+    render inline: EmberCli[name].sprockets.index_html(head: head, body: body)
   end
 
   def include_ember_script_tags(name, **options)
-    javascript_include_tag(*EmberCli[name].exposed_js_assets, options)
+    javascript_include_tag(*EmberCli[name].sprockets.assets, options)
   end
 
   def include_ember_stylesheet_tags(name, **options)
-    stylesheet_link_tag(*EmberCli[name].exposed_css_assets, options)
+    stylesheet_link_tag(*EmberCli[name].sprockets.assets, options)
   end
 end
