+++ lib/ember-cli/app.rb
@@ -1,6 +1,5 @@
-require "non-stupid-digest-assets"
-require "ember-cli/html_page"
 require "ember-cli/shell"
+require "ember-cli/sprockets"
 require "ember-cli/build_monitor"
 
 module EmberCli
@@ -8,7 +7,8 @@ module EmberCli
     attr_reader :name, :options, :paths
 
     def initialize(name, **options)
-      @name, @options = name.to_s, options
+      @name = name.to_s
+      @options = options
       @paths = PathSet.new(
         app: self,
         configuration: EmberCli.configuration,
@@ -54,27 +54,16 @@ module EmberCli
       @shell.test
     end
 
-    def index_html(sprockets:, head:, body:)
-      html_page = HtmlPage.new(
-        content: index_file.read,
-        head: head,
-        body: body,
-      )
-
-      html_page.render
+    def sprockets
+      EmberCli::Sprockets.new(self)
     end
 
-    def exposed_js_assets
-      [vendor_assets, application_assets]
-    end
-    alias exposed_css_assets exposed_js_assets
-
-    def vendor_assets
-      "#{name}/assets/vendor"
-    end
-
-    def application_assets
-      "#{name}/assets/#{ember_app_name}"
+    def index_file
+      if EmberCli.env.production?
+        paths.applications.join("#{name}.html")
+      else
+        paths.dist.join("index.html")
+      end
     end
 
     def wait
@@ -87,7 +76,7 @@ module EmberCli
       @prepared ||= begin
         @build.reset
         symlink_to_assets_root
-        add_assets_to_precompile_list
+        sprockets.register!
         true
       end
     end
@@ -98,35 +87,11 @@ module EmberCli
       end
     end
 
-    def index_file
-      if EmberCli.env.production?
-        paths.applications.join("#{name}.html")
-      else
-        paths.dist.join("index.html")
-      end
-    end
-
     def symlink_to_assets_root
       paths.app_assets.make_symlink paths.dist
     rescue Errno::EEXIST
       # Sometimes happens when starting multiple Unicorn workers.
       # Ignoring...
-    end
-
-    def add_assets_to_precompile_list
-      assets = %r{\A#{name}/}
-
-      Rails.configuration.assets.precompile << assets
-      NonStupidDigestAssets.whitelist << assets
-    end
-
-    def ember_app_name
-      @ember_app_name ||= options.fetch(:name){ package_json.fetch(:name) }
-    end
-
-    def package_json
-      @package_json ||=
-        JSON.parse(paths.package_json_file.read).with_indifferent_access
     end
 
     def excluded_ember_deps
